pyproject.toml
[tool.poetry]
name = "servier"
version = "0.1.0"
description = ""
authors = ["Lahloug <ghali-lahlou998@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.0"
pandas = "^1.2.1"
apache-airflow = "^2.0.0"
scipy = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
servier-read = "servier.read:read_drugs"
